Refactor visibility logic for entityTypes to be permissive
