+++ __tests__/__mocks__/splitUpPackagesCsn.json
