+++ __tests__/unittest/templates.test.js
@@ -317,461 +317,4 @@ describe('templates', () => {
             expect(eventResourceTemplate).toEqual([]);
         });
     });
-    describe('createEntityTypeTemplate', () => {
-        const packageIds = [
-            'sap.test.cdsrc.sample:package:test-entityType-public:v1',
-            'sap.test.cdsrc.sample:package:test-entityType-private:v1'
-        ];
-
-        it('should exclude EntityType if referenced by a private API', () => {
-            const entity = {
-                ordId: "sap.sm:entityType:PrivateEntity:v1",
-                entityName: "PrivateEntity",
-            };
-
-            const updatedAppConfig = {
-                ...appConfig,
-                apiResources: [{
-                    ordId: "sap.sm:apiResource:SomeAPI:v1",
-                    visibility: "private",
-                    entityTypeMappings: [{ entityTypeTargets: [{ ordId: entity.ordId }] }]
-                }]
-            };
-
-            const entityType = createEntityTypeTemplate(updatedAppConfig, packageIds, entity);
-
-            expect(entityType).toBeNull();
-        });
-
-
-        it('should exclude EntityType if referenced by a private DataProduct', () => {
-            const entity = {
-                ordId: "sap.sm:entityType:PrivateEntity:v1",
-                entityName: "PrivateEntity",
-            };
-
-            const updatedAppConfig = {
-                ...appConfig,
-                dataProducts: [{
-                    ordId: "sap.sm:dataProduct:SomeDataProduct:v1",
-                    visibility: "private",
-                    entityTypes: [entity.ordId]
-                }]
-            };
-
-            const entityType = createEntityTypeTemplate(updatedAppConfig, packageIds, entity);
-
-            expect(entityType).toBeNull();
-        });
-
-        it('should exclude EntityType if referenced by a private Event Resource', () => {
-            const entity = {
-                ordId: "sap.sm:entityType:PrivateEntity:v1",
-                entityName: "PrivateEntity",
-            };
-
-            const updatedAppConfig = {
-                ...appConfig,
-                eventResources: [{
-                    ordId: "sap.sm:eventResource:SomeEvent:v1",
-                    visibility: "private",
-                    entityTypeMappings: [{ entityTypeTargets: [{ ordId: entity.ordId }] }]
-                }]
-            };
-
-            const entityType = createEntityTypeTemplate(updatedAppConfig, packageIds, entity);
-
-            expect(entityType).toBeNull();
-        });
-
-    });
-
-
-    describe('createAPIResourceTemplate', () => {
-        const packageIds = [
-            'sap.test.cdsrc.sample:package:test-api-public:v1',
-            'sap.test.cdsrc.sample:package:test-api-private:v1',
-            'sap.test.cdsrc.sample:package:test-api-internal:v1'
-        ];
-
-        it('should only include referenced entity types from the correct namespace', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = {};
-            const expectedOrdId = `${appConfig.ordNamespace}:apiResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }] },
-                    { entityTypeTargets: [{ ordId: `external.namespace:entityType:InvalidEntity:v1` }] }
-                ]
-            };
-
-            const apiResource = createAPIResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(apiResource[0].ordId).toEqual(expectedOrdId)
-            expect(apiResource[0].entityTypeMappings).toBeDefined();
-            expect(apiResource[0].entityTypeMappings).toHaveLength(1);
-            expect(apiResource[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }
-            ]);
-        });
-
-        it('should mark API resource as private if a referenced entityType is private', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = {};
-            const expectedOrdId = `${appConfig.ordNamespace}:apiResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PrivateEntity:v1`, visibility: "private" }] }
-                ]
-            };
-
-            const apiResource = createAPIResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            if (apiResource.length === 0) {
-                expect(apiResource).toEqual([]);
-            } else {
-                expect(apiResource[0].ordId).toEqual(expectedOrdId);
-                expect(apiResource[0].visibility).toEqual('private');
-            }
-        });
-
-        it('should mark API resource as internal if a referenced entityType is internal but not private', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = {};
-            const expectedOrdId = `${appConfig.ordNamespace}:apiResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:InternalEntity:v1`, visibility: "internal" }] }
-                ]
-            };
-
-            const apiResource = createAPIResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            if (apiResource.length === 0) {
-                expect(apiResource).toEqual([]);
-            } else {
-                expect(apiResource[0].ordId).toEqual(expectedOrdId);
-                expect(apiResource[0].visibility).toEqual('internal');
-            }
-        });
-
-        it('should keep API resource public if no restricted entityTypes are referenced', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = {};
-            const expectedOrdId = `${appConfig.ordNamespace}:apiResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PublicEntity:v1`, visibility: "public" }] }
-                ]
-            };
-
-            const apiResource = createAPIResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(apiResource[0].ordId).toEqual(expectedOrdId)
-            expect(apiResource[0].visibility).toEqual('public');
-        });
-
-        it('should remove duplicate referenced entityTypes', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = {};
-            const expectedOrdId = `${appConfig.ordNamespace}:apiResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    {
-                        entityTypeTargets: [
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` },
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-                        ]
-                    }
-                ]
-            };
-
-            const apiResource = createAPIResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(apiResource[0].ordId).toEqual(expectedOrdId)
-            expect(apiResource[0].entityTypeMappings[0].entityTypeTargets).toHaveLength(1);
-            expect(apiResource[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-            ]);
-        });
-    });
-
-    describe('createEventResourceTemplate', () => {
-        const packageIds = [
-            'sap.test.cdsrc.sample:package:test-event-public:v1',
-            'sap.test.cdsrc.sample:package:test-event-private:v1',
-            'sap.test.cdsrc.sample:package:test-event-internal:v1'
-        ];
-
-        it('should only include referenced entity types from the correct namespace', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = linkedModel;
-            const expectedOrdId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }] },
-                    { entityTypeTargets: [{ ordId: `external.namespace:entityType:InvalidEntity:v1` }] }
-                ]
-            };
-
-            const eventResource = createEventResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-        if (eventResource.length === 0) {
-            expect(eventResource).toEqual([]);
-        } else {
-            expect(eventResource[0].ordId).toEqual(expectedOrdId);
-            expect(eventResource[0].entityTypeMappings).toBeDefined();
-            expect(eventResource[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }
-            ]);
-        }
-        });
-
-        it('should mark Event resource as private if a referenced entityType is private', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = linkedModel;
-            const expectedOrdId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PrivateEntity:v1`, visibility: "private" }] }
-                ]
-            };
-
-            const eventResource = createEventResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-        if (eventResource.length === 0) {
-            expect(eventResource).toEqual([]);
-        } else {
-            expect(eventResource[0].ordId).toEqual(expectedOrdId);
-            expect(eventResource[0].visibility).toEqual('private');
-        }
-        });
-
-        it('should mark Event resource as internal if a referenced entityType is internal but not private', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = linkedModel;
-            const expectedOrdId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:InternalEntity:v1`, visibility: "internal" }] }
-                ]
-            };
-
-            const eventResource = createEventResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(eventResource[0].ordId).toEqual(expectedOrdId);
-            expect(eventResource[0].visibility).toEqual('internal');
-        });
-
-        it('should keep Event resource public if no restricted entityTypes are referenced', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = linkedModel;
-            const expectedOrdId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PublicEntity:v1`, visibility: "public" }] }
-                ]
-            };
-
-            const eventResource = createEventResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(eventResource[0].ordId).toEqual(expectedOrdId);
-            expect(eventResource[0].visibility).toEqual('public');
-        });
-
-        it('should remove duplicate referenced entityTypes', () => {
-            const serviceName = 'MyService';
-            const serviceDefinition = linkedModel;
-            const expectedOrdId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    {
-                        entityTypeTargets: [
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` },
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-                        ]
-                    }
-                ]
-            };
-
-            const eventResource = createEventResourceTemplate(serviceName, serviceDefinition, updatedAppConfig, packageIds, {});
-
-            expect(eventResource[0].ordId).toEqual(expectedOrdId);
-            expect(eventResource[0].entityTypeMappings[0].entityTypeTargets).toHaveLength(1);
-            expect(eventResource[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-            ]);
-        });
-    });
-
-    describe('createDataProductTemplate', () => {
-        const packageIds = [
-            'sap.test.cdsrc.sample:package:test-dataProduct-public:v1',
-            'sap.test.cdsrc.sample:package:test-dataProduct-private:v1',
-            'sap.test.cdsrc.sample:package:test-dataProduct-internal:v1'
-        ];
-
-        const appConfig = {
-            ordNamespace: 'customer.testNamespace',
-            appName: 'testAppName',
-            lastUpdate: '2024-03-13',
-            entityTypeMappings: []
-        };
-
-        it('should only include referenced entity types from the correct namespace', () => {
-            const dataProductName = 'ValidDataProduct';
-            const expectedOrdId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }] },
-                    { entityTypeTargets: [{ ordId: `external.namespace:entityType:InvalidEntity:v1` }] }
-                ]
-            };
-
-            const dataProductDefinition = {
-                "@title": "Valid Data Product",
-                "@ORD.Extensions.entityTypes": [
-                    { ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }
-                ]
-            };
-
-            const dataProductTemplate = createDataProductTemplate(dataProductName, dataProductDefinition, updatedAppConfig, packageIds);
-
-            expect(dataProductTemplate[0].ordId).toEqual(expectedOrdId);
-            expect(dataProductTemplate[0].entityTypeMappings).toBeDefined();
-            expect(dataProductTemplate[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:ValidEntity:v1` }
-            ]);
-        });
-
-        it('should exclude Data Product or mark Data Product as private if a referenced entityType is private', () => {
-            const dataProductName = 'PrivateDataProduct';
-            const expectedOrdId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PrivateEntity:v1`, visibility: "private" }] }
-                ]
-            };
-
-            const dataProductDefinition = {
-                "@title": "Private Data Product",
-                "@ORD.Extensions.entityTypes": [
-                    { ordId: `${appConfig.ordNamespace}:entityType:PrivateEntity:v1` }
-                ]
-            };
-
-            const dataProductTemplate = createDataProductTemplate(dataProductName, dataProductDefinition, updatedAppConfig, packageIds);
-
-            if (dataProductTemplate.length === 0) {
-                expect(dataProductTemplate).toEqual([]);
-            } else {
-                expect(dataProductTemplate[0].ordId).toEqual(expectedOrdId);
-                expect(dataProductTemplate[0].visibility).toEqual('private');
-            }
-        });
-
-        it('should mark Data Product as internal if a referenced entityType is internal but not private', () => {
-            const dataProductName = 'InternalDataProduct';
-            const expectedOrdId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:InternalEntity:v1`, visibility: "internal" }] }
-                ]
-            };
-
-            const dataProductDefinition = {
-                "@title": "Internal Data Product",
-                "@ORD.Extensions.entityTypes": [
-                    { ordId: `${appConfig.ordNamespace}:entityType:InternalEntity:v1` }
-                ]
-            };
-
-            const dataProductTemplate = createDataProductTemplate(dataProductName, dataProductDefinition, updatedAppConfig, packageIds);
-
-            expect(dataProductTemplate[0].ordId).toEqual(expectedOrdId);
-            expect(dataProductTemplate[0].visibility).toEqual('internal');
-        });
-
-        it('should keep Data Product public if no restricted entityTypes are referenced', () => {
-            const dataProductName = 'PublicDataProduct';
-            const expectedOrdId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    { entityTypeTargets: [{ ordId: `${appConfig.ordNamespace}:entityType:PublicEntity:v1`, visibility: "public" }] }
-                ]
-            };
-
-            const dataProductDefinition = {
-                "@title": "Public Data Product",
-                "@ORD.Extensions.entityTypes": [
-                    { ordId: `${appConfig.ordNamespace}:entityType:PublicEntity:v1` }
-                ]
-            };
-
-            const dataProductTemplate = createDataProductTemplate(dataProductName, dataProductDefinition, updatedAppConfig, packageIds);
-
-            expect(dataProductTemplate[0].ordId).toEqual(expectedOrdId);
-            expect(dataProductTemplate[0].visibility).toEqual('public');
-        });
-
-        it('should remove duplicate referenced entityTypes', () => {
-            const dataProductName = 'DuplicateEntityDataProduct';
-            const expectedOrdId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
-
-            const updatedAppConfig = {
-                ...appConfig,
-                entityTypeMappings: [
-                    {
-                        entityTypeTargets: [
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` },
-                            { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-                        ]
-                    }
-                ]
-            };
-
-            const dataProductDefinition = {
-                "@title": "Duplicate Entity Data Product",
-                "@ORD.Extensions.entityTypes": [
-                    { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-                ]
-            };
-
-            const dataProductTemplate = createDataProductTemplate(dataProductName, dataProductDefinition, updatedAppConfig, packageIds);
-
-            expect(dataProductTemplate[0].ordId).toEqual(expectedOrdId);
-            expect(dataProductTemplate[0].entityTypeMappings[0].entityTypeTargets).toHaveLength(1);
-            expect(dataProductTemplate[0].entityTypeMappings[0].entityTypeTargets).toEqual([
-                { ordId: `${appConfig.ordNamespace}:entityType:DuplicateEntity:v1` }
-            ]);
-        });
-    });
 });
