+++ lib/templates.js
@@ -195,46 +195,14 @@ const createEntityTypeTemplate = (appConfig, packageIds, entity) => {
     const ordExtensions = readORDExtensions(entity);
     const visibility = ordExtensions.visibility || RESOURCE_VISIBILITY.public;
 
-    // Check whether the EntityType is referenced by a restrictive API
-    const referencedByRestrictedAPI = (Array.isArray(appConfig.apiResources) ? appConfig.apiResources : []).some(api =>
-        api.entityTypeMappings?.some(mapping =>
-            mapping.entityTypeTargets.some(target => target.ordId === entity.ordId)
-        ) && [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(api.visibility)
-    );
-
-   // Check whether the EntityType is referenced by a restrictive event resource
-    const referencedByRestrictedEvent = (Array.isArray(appConfig.eventResources) ? appConfig.eventResources : []).some(event =>
-        event.entityTypeMappings?.some(mapping =>
-            mapping.entityTypeTargets.some(target => target.ordId === entity.ordId)
-        ) && [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(event.visibility)
-    );
-
-    // Check whether the EntityType is referenced by a restrictive DataProduct
-    const referencedByRestrictedDataProduct = (Array.isArray(appConfig.dataProducts) ? appConfig.dataProducts : []).some(dp =>
-        dp.entityTypes?.some(target => target === entity.ordId)
-        && [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(dp.visibility)
-    );
-
-    // Adopt the most restrictive visibility
-    let finalVisibility = visibility;
-    if (referencedByRestrictedAPI || referencedByRestrictedEvent || referencedByRestrictedDataProduct) {
-        if (referencedByRestrictedAPI && appConfig.apiResources.some(api => api.visibility === RESOURCE_VISIBILITY.private) ||
-            referencedByRestrictedEvent && appConfig.eventResources.some(event => event.visibility === RESOURCE_VISIBILITY.private) ||
-            referencedByRestrictedDataProduct && appConfig.dataProducts.some(dp => dp.visibility === RESOURCE_VISIBILITY.private)) {
-            finalVisibility = RESOURCE_VISIBILITY.private;
-        } else {
-            finalVisibility = RESOURCE_VISIBILITY.internal;
-        }
-    }
-
-    if (finalVisibility === RESOURCE_VISIBILITY.private) {
+    if (visibility === RESOURCE_VISIBILITY.private) {
         return null;
     }
 
     const packageId = packageIds.find(id =>
-        finalVisibility === RESOURCE_VISIBILITY.public
+        visibility === RESOURCE_VISIBILITY.public
             ? id.includes(`entityType`) && (!id.includes("-internal") && !id.includes("-private"))
-            : id.includes(`entityType-${finalVisibility}`)
+            : id.includes(`entityType-${visibility}`)
     );
 
     return {
@@ -245,7 +213,7 @@ const createEntityTypeTemplate = (appConfig, packageIds, entity) => {
         description: DESCRIPTION_PREFIX + entity.entityName,
         version: _getEntityVersion(entity),
         lastUpdate: appConfig.lastUpdate,
-        visibility: finalVisibility,
+        visibility,
         partOfPackage: packageId || _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.entityType),
         releaseStatus: "active",
         level: entity["@ObjectModel.compositionRoot"] || entity["@ODM.root"] ? LEVEL.rootEntity : LEVEL.subEntity,
@@ -289,41 +257,6 @@ const createAPIResourceTemplate = (serviceName, serviceDefinition, appConfig, pa
             resourceDefinitions.push(_getResourceDefinition("edmx", "xml", ordId, serviceName, "edmx", accessStrategies));
         }
 
-        // Ensure API resource respects referenced EntityType visibility
-        const referencedEntityTypes = Array.isArray(appConfig.entityTypeMappings)
-            ? appConfig.entityTypeMappings.flatMap(mapping => mapping.entityTypeTargets)
-                .filter(target => target.ordId.startsWith(appConfig.ordNamespace))
-            : [];
-
-        const hasRestrictedEntityType = Array.isArray(appConfig.entityTypeMappings)
-            ? appConfig.entityTypeMappings.some(mapping =>
-                mapping.entityTypeTargets.some(target =>
-                    referencedEntityTypes.some(et => et.ordId === target.ordId) &&
-                    [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(target.visibility)
-                )
-            )
-            : false;
-
-            let finalVisibility;
-
-            if (hasRestrictedEntityType) {
-                if (referencedEntityTypes.some(et => et.visibility === RESOURCE_VISIBILITY.private)) {
-                    finalVisibility = RESOURCE_VISIBILITY.private;
-                } else {
-                    finalVisibility = RESOURCE_VISIBILITY.internal;
-                }
-            } else {
-                finalVisibility = visibility;
-            }
-            // Make sure that the EntityType package is also “internal” if an associated API is internal
-            const entityTypePackageId = referencedEntityTypes.length > 0
-            ? packageIds.find(id =>
-                finalVisibility === RESOURCE_VISIBILITY.public
-                    ? id.includes("entityType") && !id.includes("-internal") && !id.includes("-private")
-                    : id.includes(`entityType-internal`) // If the API is internal, entityType is also internal
-            )
-            : null;
-
         let obj = {
             ordId,
             title:
@@ -336,7 +269,7 @@ const createAPIResourceTemplate = (serviceName, serviceDefinition, appConfig, pa
                 DESCRIPTION_PREFIX + serviceName,
             version: "1.0.0",
             lastUpdate: appConfig.lastUpdate,
-            visibility: finalVisibility,
+            visibility,
             partOfPackage: packageId || _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.api),
             partOfGroups: [_getGroupID(serviceName, defaults.groupTypeId, appConfig)],
             releaseStatus: "active",
@@ -348,15 +281,16 @@ const createAPIResourceTemplate = (serviceName, serviceDefinition, appConfig, pa
                 supported: "no",
             },
             integrationDependency: Array.isArray(packageIds)
-            ? packageIds.find(id =>
-                id.includes("integrationDependency") &&
-                (finalVisibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
-            ) || null
-            : null,
-            entityTypeMappings: referencedEntityTypes.length > 0
+                ? packageIds.find(id =>
+                    id.includes("integrationDependency") &&
+                    (visibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
+                ) || null
+                : null,
+            entityTypeMappings: Array.isArray(appConfig.entityTypeMappings)
                 ? [{
-                    entityTypeTargets: _.uniqBy(referencedEntityTypes.map(et => ({ "ordId": et.ordId })), 'ordId'),
-                    partOfPackage: entityTypePackageId
+                    entityTypeTargets: _.uniqBy(
+                        appConfig.entityTypeMappings.flatMap(m => m.entityTypeTargets).map(et => ({ ordId: et.ordId })), 'ordId'
+                    ),
                 }]
                 : [],
             ...ordExtensions,
@@ -395,47 +329,7 @@ const createEventResourceTemplate = (serviceName, serviceDefinition, appConfig, 
     : null;
 
     const ordId = `${appConfig.ordNamespace}:eventResource:${serviceName}:v1`;
-
-      // Ensure Event resource respects referenced EntityType visibility
-    const referencedEntityTypes = Array.isArray(appConfig.entityTypeMappings)
-    ? appConfig.entityTypeMappings.flatMap(mapping => mapping.entityTypeTargets)
-        .filter(target => target.ordId.startsWith(appConfig.ordNamespace))
-    : [];
-
-    const hasRestrictedEntityType = Array.isArray(appConfig.entityTypeMappings)
-        ? appConfig.entityTypeMappings.some(mapping =>
-            mapping.entityTypeTargets.some(target =>
-                referencedEntityTypes.some(et => et.ordId === target.ordId) &&
-                [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(target.visibility)
-            )
-        )
-        : false;
-
-    let finalVisibility;
-
-    if (hasRestrictedEntityType) {
-        if (referencedEntityTypes.some(et => et.visibility === RESOURCE_VISIBILITY.private)) {
-            finalVisibility = RESOURCE_VISIBILITY.private;
-        } else {
-            finalVisibility = RESOURCE_VISIBILITY.internal;
-        }
-    } else {
-        finalVisibility = visibility;
-    }
-
-    if (finalVisibility === RESOURCE_VISIBILITY.private) {
-        return [];
-    }
-
-    const entityTypePackageId = referencedEntityTypes.length > 0
-    ? packageIds.find(id =>
-        finalVisibility === RESOURCE_VISIBILITY.public
-            ? id.includes("entityType") && !id.includes("-internal") && !id.includes("-private")
-            : id.includes("entityType-internal") // Falls das Event internal ist, wird auch das entityType-Paket internal
-    )
-    : null;
-
-    return [{
+    let obj = {
         ordId,
         title:
             serviceDefinition["@title"] ??
@@ -443,30 +337,34 @@ const createEventResourceTemplate = (serviceName, serviceDefinition, appConfig, 
             `ODM ${appConfig.appName.replace(/[^a-zA-Z0-9]/g, "")} Events`,
         shortDescription: `${serviceName} event resource`,
         description:
-            serviceDefinition['@description'] ?? serviceDefinition['@Core.Description'] ??
+            serviceDefinition["@description"] ??
+            serviceDefinition["@Core.Description"] ??
             "CAP Event resource describing events / messages.",
         version: "1.0.0",
         lastUpdate: appConfig.lastUpdate,
         releaseStatus: "active",
         partOfPackage: packageId || _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.event),
         partOfGroups: [_getGroupID(serviceName, defaults.groupTypeId, appConfig)],
-        visibility: finalVisibility,
+        visibility,
         resourceDefinitions: [_getResourceDefinition("asyncapi-v2", "json", ordId, serviceName, "asyncapi2.json", accessStrategies)],
         extensible: { supported: "no" },
         integrationDependency: Array.isArray(packageIds)
             ? packageIds.find(id =>
                 id.includes("integrationDependency") &&
-                (finalVisibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
+                (visibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
             ) || null
             : null,
-        entityTypeMappings: referencedEntityTypes.length > 0
+        entityTypeMappings: Array.isArray(appConfig.entityTypeMappings)
             ? [{
-                entityTypeTargets: _.uniqBy(referencedEntityTypes.map(et => ({ "ordId": et.ordId })), 'ordId'),
-                partOfPackage: entityTypePackageId // Korrekte Paketzuweisung für EntityTypes
+                entityTypeTargets: _.uniqBy(
+                    appConfig.entityTypeMappings.flatMap(m => m.entityTypeTargets).map(et => ({ ordId: et.ordId })), 'ordId'
+                ),
             }]
             : [],
         ...ordExtensions
-    }]
+    };
+
+    return obj.visibility === RESOURCE_VISIBILITY.public || obj.visibility === RESOURCE_VISIBILITY.internal ? [obj] : [];
 };
 
 /**
@@ -483,82 +381,46 @@ const createDataProductTemplate = (dataProductName, dataProductDefinition, appCo
     const ordExtensions = readORDExtensions(dataProductDefinition);
     const visibility = ordExtensions.visibility || RESOURCE_VISIBILITY.public;
 
-    const referencedEntityTypes = Array.isArray(appConfig.entityTypeMappings)
-    ? appConfig.entityTypeMappings.flatMap(mapping => mapping.entityTypeTargets)
-        .filter(target => target.ordId.startsWith(appConfig.ordNamespace))
-    : [];
-
-    const hasRestrictedEntityType = Array.isArray(appConfig.entityTypeMappings)
-    ? appConfig.entityTypeMappings.some(mapping =>
-        mapping.entityTypeTargets.some(target =>
-            referencedEntityTypes.some(et => et.ordId === target.ordId) &&
-            [RESOURCE_VISIBILITY.private, RESOURCE_VISIBILITY.internal].includes(target.visibility)
-        )
-    )
-    : false;
-
-    let finalVisibility;
-
-    if (hasRestrictedEntityType) {
-        if (referencedEntityTypes.some(et => et.visibility === RESOURCE_VISIBILITY.private)) {
-            finalVisibility = RESOURCE_VISIBILITY.private;
-        } else {
-            finalVisibility = RESOURCE_VISIBILITY.internal;
-        }
-    } else {
-        finalVisibility = visibility;
-    }
-
-    if (finalVisibility === RESOURCE_VISIBILITY.private) {
-        return [];
-    }
-
     const packageId = Array.isArray(packageIds)
         ? packageIds.find(id =>
-            finalVisibility === RESOURCE_VISIBILITY.public
+            visibility === RESOURCE_VISIBILITY.public
                 ? id.includes("dataProduct") && !id.includes("-internal") && !id.includes("-private")
-                : id.includes(`dataProduct-${finalVisibility}`)
+                : id.includes(`dataProduct-${visibility}`)
         )
         : null;
 
     const ordId = `${appConfig.ordNamespace}:dataProduct:${dataProductName}:v1`;
 
-    const entityTypePackageId = referencedEntityTypes.length > 0
-    ? packageIds.find(id =>
-        finalVisibility === RESOURCE_VISIBILITY.public
-            ? id.includes("entityType") && !id.includes("-internal") && !id.includes("-private")
-            : id.includes("entityType-internal") // Falls das Data Product internal ist, wird auch das entityType-Paket internal
-    )
-    : null;
-
-    return [{
+    let obj = {
         ordId,
         title: dataProductDefinition["@title"] ?? dataProductName,
         shortDescription: dataProductDefinition["@shortDescription"] ?? `${dataProductName} Data Product`,
         description: dataProductDefinition["@description"] ?? `Data Product for ${dataProductName}.`,
         version: "1.0.0",
         lastUpdate: appConfig.lastUpdate,
-        visibility: finalVisibility,
+        visibility,
         partOfPackage: packageId || _getPackageID(appConfig.ordNamespace, packageIds, ORD_RESOURCE_TYPE.dataProduct),
         releaseStatus: "active",
         type: "primary",
         category: "business-object",
         outputPorts: ordExtensions.outputPorts || [],
-        entityTypeMappings: referencedEntityTypes.length > 0
+        entityTypeMappings: Array.isArray(appConfig.entityTypeMappings)
             ? [{
-                entityTypeTargets: _.uniqBy(referencedEntityTypes.map(et => ({ "ordId": et.ordId })), 'ordId'),
-                partOfPackage: entityTypePackageId // Korrekte Paketzuweisung für EntityTypes
+                entityTypeTargets: _.uniqBy(
+                    appConfig.entityTypeMappings.flatMap(m => m.entityTypeTargets).map(et => ({ ordId: et.ordId })), 'ordId'
+                ),
             }]
             : [],
         integrationDependency: Array.isArray(packageIds)
             ? packageIds.find(id =>
                 id.includes("integrationDependency") &&
-                (finalVisibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
+                (visibility === RESOURCE_VISIBILITY.public ? !id.includes("-internal") : id.includes("-internal"))
             ) || null
             : null,
-
         ...ordExtensions,
-    }];
+    };
+
+    return obj.visibility === RESOURCE_VISIBILITY.public || obj.visibility === RESOURCE_VISIBILITY.internal ? [obj] : [];
 };
 
 function _getPackageID(namespace, packageIds, resourceType) {
